Add Identity Credential HAL, default implementation, and VTS tests.
IIdentityCredentialStore provides an interface to a secure store for
user identity documents.  This HAL is deliberately fairly general and
abstract.  To the extent possible, specification of the message
formats and semantics of communication with credential verification
devices and issuing authorities (IAs) is out of scope for this HAL.

It provides the interface with secure storage but a
credential-specific Android application will be required to implement
the presentation and verification protocols and processes appropriate
for the specific credential type.

TODOs before landing:

 - Need to specify where somewhere in the Attestation Data that the
   created key is a CredentialKey and not a regular Android Keystore
   key. This should be done in a way similar to how a Protected
   Confirmation conveys in its attestation data that it has the
   TRUSTED_CONFIRMATION_REQUIRED flag set.

 - Expand VTS tests - base on existing CTS tests for credstore.

 - Go through and resolve all TODOs in code.

Bug: 111446262
Test: VtsHalIdentityCredentialV1_0TargetTest
Change-Id: I64eb50114d645dd475012ad1b889d2177aaf1d37