Update patch set 2
Patch Set 2: Code-Review-2

hmmmmmm... Sorry, although I understand this makes your life easier, composer 2.2 and 2.3 are not supported in adapter and we won't support them. Adapter is supposed to be used when we moved to HAL and developed composer 2.1, it's not supposed to be used when upgrading to newer composer version.

Patch-set: 2
Reviewer: Gerrit User 1267775 <1267775@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2