Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is also blocking the test suite run in the lab, due to it's significant load.

I would like more information on this CL if possible. Which branch and which device is this running long on? There used to be a problem with VtsMultiDeviceTest's ProcessHelper; specifically, the binary gtest would take ~1 hour to run, but ProcessHelper would take ~12 hours to parse and print the results. My understanding is once we switched to gtest, the tests only took ~1 hour end-to-end, so the problem would be resolved.

Patch-set: 1
Reviewer: Gerrit User 1125923 <1125923@85c56323-6fa9-3386-8a01-6480fb634889>