Remove creation time checks.
We should not be relying on the HAL service to add CREATION_TIME to
keys.  It was always intended to be an optional tag that could be
added by the caller.  One widespread implementation started adding
it (somewhat erroneously) if it wasn't provided, and it appears that
this implementation's behavior became assumed to be the required
behavior.

Test: VtsHalKeymasterV4_0TargetTest
Change-Id: I34267c4e1f59fd8ee5f898f8c746a7b49f4d74a5