Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is exponent 3 explicitly banned?

No, but this follows in the same vein as the modifications to HMAC key size. If we test more than what is explicitly required, then we're just adding implicit requirements on top of what is laid out in the HAL. This has led to confusion from partners.

Patch-set: 1