Adding tests for INCLUE_UNIQUE_ID
The tests in this patch check that the behavior of INCLUDE_UNIQUE_ID in
key attestations is behaving as expected. The primary behaviors to test
are that
(1)keys in the same 30 day time window with the same application id
   have the same unique ID in the attestation certificate.
(2)keys in different 30 day time windows with the same application id
   have different unique IDs in the attestation certificate.
(3)keys in the same 30 day time window with the same application id have
  the same unique ID unless RESET_SINCE_ID_ROTATION is present in the
  authorization list. This implies that a factory reset has occurred.

By testing that the CREATION_DATETIME recorded in the certificate
matches what was supplied by the HAL caller, this test also checks that
Keymaster implementations accept HAL caller provided CREATION_DATETIME
values.

Bug: 143943044
Test: atest VtsHalKeymasterV4_0TargetTest
Change-Id: I82331884e55466e7b3f295f36d290dd985a9606e