Add SCOPED_TRACE to disambiguate flush failures
Since all sensors are tested in the flush test in parallel, it can be
difficult to tell which sensor is triggering the test to fail. Add
SCOPED_TRACE instances in runFlushTest so that the failure message will
be accompanied by identifiers for the failing sensor.

Also, convert the ASSERT_EQ check for validating the number of flush
complete events received to EXPECT_EQ, so error messages will be printed
for all failing sensors, rather than only the first one to fail the
check.

Test: TODO
Change-Id: I428bf7aa476ffebba3279c2524958055101827d4