Update patch set 2
Patch Set 2:

> > I believe this should work without the framework being enabled and my apologies for the oversight. I'm reluctant to enable the framework and native service as not all HAL implementations support multiple clients. This is likely to cause grief with at least some other device out there that would be unhappy if we enabled native service and test harness as HAL clients concurrently.
> 
> The property maybe not the ideal way. see here https://android-review.googlesource.com/c/platform/test/vts-testcase/hal/+/1157473
> The Intel engineer said in their device they do not have "the property" but have "the feature". FYI, I test it in Volvo device. Volvo has both.

Thanks for the info Kai. It may not be ideal, but this is extremely simple to add for Intel. More importantly, Volvo already has it and since we can't use pm, using features is not an option. This is at least a viable alternative imo.

Patch-set: 2
Reviewer: Gerrit User 1315593 <1315593@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0