Add CellIdentity to getBarringInfoResponse
Add CellIdentity to getBarringInfoResponse to make
the req/resp method match the existing indication.
The CellIdentity is used to prevent races where
barring is attributed to the wrong cell based on the
timing of separate polled or unsolicited inputs to
the framework.

Because the CellIdentity is already in IRadioIndication
the information to provide CID with BarringInfo is already
present and available. This just makes the unsol and the
polled APIs match.

Matching indication is IRadioIndication@1.5::barringInfoChanged

Bug: 148992393
Test: make && VtsHalRadioV1_5TargetTest
Merged-In: I74e6ab9ca9e0dd49846af774fd73773ff3737df7
Change-Id: I74e6ab9ca9e0dd49846af774fd73773ff3737df7
(cherry picked from commit 6894f89fd88dab3b56531c06e8f5d2ae0316f47f)