Update patch set 5
Patch Set 5:

> I commented it out instead of remove them to another CL because I tend have git
> problems and don't want to spend a long time untangle any mistake in the process.
> I will remove the commented out codes before we submit and merge in the codes.

We're at the point where we're ready to merge the attestation validation code so please remove it now. We can circle back and add it later.

(No need to be concerned that the code is going to be deleted - for example, it's already here in Gerrit so you can always get it from there.)

Patch-set: 5
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>