Identity: factor out crypto in a small C library with abstraction layer.
All crypto used in Identity Credential is moved into a small C library
(currently just EmbeddedIc.[ch]) which has no dependencies, not even
the standard C library. Operations actually needed (such as crypto,
etc.) are defined as eicOps* functions and is to be provided by the
platform.

Additionally, this C library has guaranteed bounded memory usage and
it does not need any heap or allocations.

The idea here is that the Secure Area would easily be able to
integrate this small C library and tie it together with its own native
crypto functions and platform specific stuff such as attestation. The
Identity Credential specifics - including generating CBOR - would be
hidden inside this library. Assuming the Secure Area already has a
KeyMaster implementation, integrating this library should be orders of
magnitudes simpler than implementing Identity Credential from scratch.

For the default implementation of Identity Credential which runs
entirely inside Android, the eicOps* operations are currently provided
by EicAndroidOps.cpp which essentially just calls into BoringSSL and
the standard C and C++ libraries. This CL includes usage in
WritableIdentityCredential.cpp just for testing.. a real
implementation would use RPC to the Secure Area which embeds this
library.

This model mimics what is successfully already used in AVB, see

 https://android.googlesource.com/platform/external/avb/+/master/README.md

for more details.

NOTE: This is just an early proof of concept... currently all we do is
to calculate the SHA-256 for the ProofOfProvisioning CBOR during
personalization. We still need tons of other stuff such as key
generation, signing, etc... but so far it's promising insofar we can
generate the SHA-256 for any CBOR of any size using very little RAM
(4-5 KiB).

Test: VTS/CTS tests pass
Bug: None
Change-Id: I3bf43fa7fd9362f94023052591801f2094a04607