Update patch set 3
Patch Set 3:

Here's an update with a lot of changes. Improvements

 - No allocation is ever performed and all structs can be allocated on the stack
 - Ops are no longer in a struct with function pointer.
 - Much lower memory usage.

For provisioning (presentation is similar but let's tackle provisioning first), we need to keep state around and this is currently stored in EicProvisioning. Notes

 - The main part of this is the state for the SHA-256 hasher which on the Android implementation is 112 bytes (BoringSSL). For Citadel, I understand this could maybe be zero?

 - In addition to this, we need to store entry count for namespaces which with a max of 32 namespaces (a reasonable limit). With a maximum of 256 entries per namespace, this comes out to 32 bytes.

 - We also need to store access control profile ids and a reasonable limit for this is 32. We probably
   can limit the identifier to be a 8-bit value in which case we need 32 bytes.

With other bookkepping etc. this brings the size of EicProvision down to 216 bytes of which 112 bytes are the digester. I don't think we can do much better than that... except that these 104 bytes actually can be serialized so presumably you can send it back to Android and ask Android to send it back in the next RPC. You'd need to guarantee that the digester is serialized too but that's another problem.. if you can do this, you actually need 0 bytes of memory!

Now that memory is under control (somewhat), I'll start focusing on implementing on adding the remaining features (MAC for Access Control Profiles, encrypting against storage key etc.) ... I expect memory usage to grow a bit (we need storageKey + friends around) but not by _that_ much.

Patch-set: 3