Update patch set 2
Patch Set 2:

Thanks for the comments.

> Patch Set 2:
> Core/vendor variants/platform/tests/etc shared linking instead of static linking will only create extra cost to load these as shared libraries.

The library is already in the zygote so there is already a copy of the functions in memory. It is becoming a public library so it is only loaded once. Yes, there's some tiny overhead in dynamic linking the 20 or so methods here. But against this, there is no need to initialize LNH internals more than once with this change, e.g. there is a JniConstants cache that will no longer be initialized once per process.

> I can't find a motivation for this in the bug. What is the benefit for avoiding static linkage in platform code?

libnativehelper is in the ART Module. The code there is updatable. If code has statically linked and the module updates that could be a compatibility risk for Mainline - it's really going to depend on what parts of the API are used.

The problem with LNH is that it is an awkward mix of stuff that doesn't really belong together:

1) JNI Registration - dirt simple, widely used.
2) JNI helper methods for public API - quite simple, come caching of JNI fields, classes, methods, etc, but fine.
3) JNI helpers for private API - yuk, problematic, doesn't mesh well with hiddenapi/core platform api. Essential in a few places, e.g. graphics and network stack.
4) JNI Invocation - intended only for android runtime and dalvikvm.

And we're also adding NDK API to it which intersects with 3) here.

> I understand the benefit for apps, but I think I am missing things for this usage.
> 
> For instance, libbinder has no static variant because the way it communicates with binder driver must be consistent everywhere, since there is no stable protocol. 

Sorry, the point being raised here is too subtle, I'm not groking it :-(

lnh is becoming a public shared library and losing the static variant. libbinder doesn't have an explicit dependency on LNH. It doesn't appear to have a transitive dependency on the dynamic variant and doesn't appear to contain any symbols from LNH.

As shown by the number of changes in this topic, there are very few users of the static variant of libnativehelper and some folks who depend on lnh when they don't need to at all.

> For instance, are we worried about ODR violations from prebuilts statically including this? Anyway, let me know. I think it might save you some work as well, or I may be missing something

ODR for LNH is a potential problem today where we can produce binaries that have static and dynamic linkage to lnh (and lnh_compat_libc++). Going forward there should be one LNH library that is used everywhere. 

OTOMH, the only parts of LNH that I'm immediately concerned about ODR violation would be the JniInvocation interface and code statically linking should not be using that anyway.

> Core/vendor variants/platform/tests/etc shared linking instead of static linking will only create extra cost to load these as shared libraries.

A tiny library that is already in memory...it's in the weeds. If not the system health team will chase us for it. As mentioned earlier we should no longer be initializing the LNH JNI constant pool per process, this is likely the same order of magnitude.

Another approach we considered was splitting out some of the "tricky" parts of LNH that are going into the NDK into a separate SO, but this (quite reasonably) gets system health push back because they don't want another SO in memory.

To summarize:

1) Less opportunity for ODR violations.
2) More robust for ART Module updates.
3) Increased time possibly because of linker costs from being only a shared library, offset by not initializing internal state in LNH for every process using the library.
4) Not a dramatic change, hardly any dependencies on static lnh (see topic span).
5) Life is simpler. Ideally, we get to just one LNH library.

Does this help?

Patch-set: 2