Update patch set 19
Patch Set 19:

With the latest PS, reader authentication is partially working.

One observation is the new eicOps() functions needed by the platform to support this, in particular the ones dealing with X.509 certificates.

Also, we currently each of the reader certificates in the chain one by one... this may need to be chunked up since each certificate is easily more than 300 bytes long.

Patch-set: 19