Update patch set 1
Patch Set 1:

I don't recall it being this way before on Android 9, perhaps we just missed the error entirely or the interface was somehow not even probed (eg. we didn't include the passthrough impl at all?).

Anyway such undefined behaviour should IMO be fixed even if we are realistically obliged to provide a thermal HAL.

Patch-set: 1