Update patch set 1
Patch Set 1:

This is an older interface that doesn't seem to be built right now. But I ended up building because of something else I'm working on, and it had the link errors described in the bug. Not sure if we want to fix this or find someway to exclude this module.

Patch-set: 1