Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-2
> > > 
> > > Changes are not allowed on HALs that are already frozen. Please file an issue and talk to our bluetooth team (jpawlowski, FYI) to do create a bluetoot.a2dp@1.1 HAL. 
> > > 
> > > I briefly looked at the HAL definition, and it looks like IBluetoothAudioOffload.startSession will need to be upreved.
> > 
> > Hi Yifan,
> > 
> > Thanks for the reply.
> > 
> > I'm going to make a v1.1, while I'm doing this, I have a question.
> > I don't get why we need to change startSession() here? 
> > Cause it seems that we have no problem passing CodecConfiguration.
> > Could you please explain more?
> > 
> > Thanks!
> 
> Oh..I think this is what you are trying to say:
> ERROR: Cannot enforce minor version uprevs for android.hardware.bluetooth.a2dp@1.1

I am not suggesting to "change" startSession, but to add a new one for 1.1 HAL with the name startSession_1_1, with the new structs that contains new enums.

Please work with our bluetooth team to get this fully supported. I don't know whether they have some other plans on this specific HAL in parallel which may conflict with your work.

Patch-set: 1