Convert Android.mk to Android.bp for vendor snapshot.
This change enables the converted modules to be included
in the vendor snapshot, which considers only
soong-defined modules.

Conversion of the modules in this directory to
Android.bp was blocked by a build-time conditional check
of TARGET_ENABLE_MEDIADRM_64. The bugs associated with
this check appear to be resolved, but there may still be
targets that do not set TARGET_ENABLE_MEDIADRM_64.

Notably:
* build/make/target/board/BoardConfigGsiCommon.mk
* device/generic/common/mgsi/BoardConfigMgsiCommon.mk

Bug: 161571531
Test: Verify inclusion of modules in vendor snapshot.
Change-Id: I47c3b739dfafdf619f8a9a85de560652966af290