Add setAllowedNetworkTypeBitmap
There are currently two methods that cover the network family types:
setPreferredNetworkType and setPreferredNetworkTypeBitmap. They both
serve a similar purpose and have been previously interpreted as a
suggestion to the modem on the type of networks requested by the OS.

This change introduces a new method (setAllowedNetworkTypeBitmap) which
will coexist with these methods. This new method must be interpreted as
a strict set of Radio Access Technology (RAT) that should be accepted by
the modem.

For instance, one may first use setAllowedNetworkTypeBitmap to exclude
2G technology from the list of usable RAT. Then, they may use
setPreferredNetworkType to describe their preference for 4G.

Test: mm
Change-Id: I7f5fdf9bf5a80c47e9ad838b740b2d6e40ffea60