Handle multi-display in VtsHalGraphicsComposer* tests
The composer tests disable vsync *only on the primary display*
via calling

mComposerClient->setVsyncEnabled(mPrimaryDisplay, false);

in the test SetUp() methods. However, these tests then call

mComposerCallback->setVsyncAllowed(false);

where the setVsyncAllowed() and onVsync() methods assume
that the vsync disablement applies to *all displays*.

This changes updates the setVsyncAllowed() method to take a
display as a parameter and updates onVsync() to check if vsync
was disabled for that particular display.

Bug: b/171305898
Test: launch Cuttlefish w/ 2 displays &&
Test: vts -m VtsHalGraphicsComposerV2_2TargetTest
Change-Id: Icac587ac9f472b81c840c11b2e70501891c855f2