Update comments for setLayerBuffer
In the spec for the setLayerBuffer function, it is stated that
no operation is performed when the layer's type is Composition::CLIENT.
However, if the composition mode of the composer is periodically changed
via "service call SurfaceFlinger 1008 i32 {0 | 1}" command, this is
called for the Composition::CLIENT type layer with new buffer_handle_t.
And in this situation, the buffer of the corresponding layer must be
updated to be synchronized with the buffer for the layer that SurfaceFlinger has.

Change-Id: I208c41fd41e91f5810c46ae6a0eacc84ded88c0f
Signed-off-by: Choi Sungyong <sywow.choi@samsung.com>