keymint: Correct the expected return on creating EARLY_BOOT_ONLY keys
Because VTS tests are run on fully-booted machines, we can only run
negative tests for early boot keys, which cannot be created or used
after /data is mounted.  This is the only test we can run in the normal
case case. The positive test will have to be done by the Android system,
when it creates/uses early boot keys during boot.  It should fail to
boot if the early boot key usage fails.

Bug: 161929178
Change-Id: Id5c8184d94474860089f530c2b483fe047a8092f