Keymaster: Fix the attestation issue in VtsHalKeymasterV3_0TargetTest
currently, by default the system/keymaster will build the verify_boot_hash
into the KM_ROOT_OF_TRUST, however, in the test case VtsHalKeymasterV3_0TargetTest
it not build it into the sequence, so the record is parsing failed in
parse_attestation_record(), and lead some attestation cases failure.

Also the default att_attestation_version is upgraded to 4, the check in the case
need to be updated.

Test: VtsHalKeymasterV3_0TargetTest
Signed-off-by: Yan, Shaopu <shaopu.yan@intel.com>
Change-Id: I3afdf83e7a1adce4d86f3980c12f5eb4674ae641