identity: Add support for encrypted provisioning.
Proof-of-concept implementation with a few TODOs sprinkled across the
code and missing documentation. Roughly the idea is that the issuer
can send an EC key and then both the Secure Area and the the issuer
use key agreement and a KDF to derive an AES-128 encryption key.

This derived encryption key is used to encrypt data element values and
decryption happens as part of provisioning. For data elements for
which this happen, ProofOfProvisioning will have an `encryptedValue`
field instead of `value` which has the encrypted data element that was
sent.

Bug: None
Test: New VTS tests which tests all of this.
Test: atest VtsHalIdentityTargetTest
Change-Id: I5ad4cd8b5b9593db032bba7e547502daea27c5c4