freeze compat matrix 6 for Android S
The current matrix is left in this same CL because this way, downstream
modifications to the current CL will automatically get absorbed into the
level 7 compatibility matrix. A CL on top of this will disable the 7
matrix and then get reverted in the aosp/master..goog/master merge
path.

WARNING: this is not actually the Android S compatibility matrix in this
CL, but rather the AOSP current CL copied into
compatibility_matrix.6.xml. The actual S compatibility matrix is not
released on AOSP. However, this is the parts of it which are available
on AOSP, added here in order to allow Android T development in a way
that will merge happily downstream.

Bug: 178221726
Test: boot device, vts_treble_vintf_vendor_test
Test: inspect matrix by diffing matrix with old matrix using:
    m analyze_matrix &&
    system/libvintf/analyze_matrix/hals_for_release.py
Change-Id: Id83986fc5089eefc2292f0042753f739f4e01a44
(cherry picked from commit 86a518cfa64fdc85ab349701e79f8adf7efdacd2)
Merged-In: Id83986fc5089eefc2292f0042753f739f4e01a44