broadcastradio@2.0 VTS uses DAB frequency invalid in some regions
DAB frequency 178352kHz hardcoded in the test is used in EU region
(channel 5C) but it is not valid in others (e.g. Korea and China).

Therefore DabTune test was modified to use one of supported
frequencies reported via getDabRegionConfig method.

Test: run VtsHalBroadcastradioV2_0TargetTest on hardware units
configured for Europe, China and Korea DAB tuner regions.

Change-Id: Icd71e42a9242528465ac1358465d8b6d6d4896bf