Make sure at least 1 SvInfo is available in TestGnssSvInfoFields
Before getting the last value from the SvInfo list, make sure at least
one SvInfo has been received, so that the list is not empty.

The issue can arise because the test makes the assumption that receiving
first location means svstatus has already been received, which is not
always true, depending on the Gnss service implementation. As an
example, in case the test is run when the receiver has already a fix,
the first location will be sent right away, and thus may be sent even
before the first svstatus (since order is not specified, afaik).

Test: run TestGnssSvInfoFields twice in a row with a Gnss service impl that sends location before svstatus
Change-Id: I70b4f39da1fdb5adf3e0e0510227686310eb6c21
Signed-off-by: Jean Dao <jean.dao@u-blox.com>