Audio HAL: Allow setters to not be implemented
setSampleRate, setChannelMask, setFormat
may not be implemented by the HAL, although this is not documented in
the HAL API.
Currently the VTS test requires their implementation if the respective
getSupported{SampleRate,ChannelMask,Format} are supported.

Relax this requirement as the framework never calls those setters.

Note that the optionality of those functions will be documented
in the next HAL API version.

Test: run vts -m VtsHalAudioV7_0Target
Bug: 69811500
Change-Id: Id0505bfb5d4812dd4c5d31a6e9d72c4c9c0cffa2