Properly convert minDelay/maxDelay (us) to ns before use in calls to batch
The minDelay/maxDelay fields of the SensorInfo struct are in
microseconds, and can be >0 for sensors with continuous reporting mode
(e.g. accelerometer, gyroscope). In multiple VTS tests, they are used as
the samplingPeriodNs argument of the batch function, which is specified
in nanoseconds. Make sure to convert the value from microseconds to
nanoseconds in those instances, otherwise the value will be rejected as
too small and the call to batch will return BAD_VALUE.

Test: run SensorsHidlTest#Batch with sensor with minDelay/maxDelay > 0 (continuous reporting mode)
Change-Id: I868cfe0140f95e471f8058f4a284aefa0110c2dd
Signed-off-by: Jean Dao <jean.dao@u-blox.com>