Prototype a Hypervisor HAL
In cases where Android is running in a virtualized scenario, it is
helpful to be able to infer certain details about the hypervisor
that is currently in use.

In some scenarios (e.g. automotive), the host/hypervisor layer is
a security and safety critical environment, and very limited and
controlled interaction is optimal when initiated by the less-secure
instance. Other environments may be more liberal in what they allow or
consider permissible interactions.

This initial HAL definition errs on the side of caution, strictly
limiting discoverability to a descriptive version information string.

Test: build
Bug: 178749018
Change-Id: Ib9f8399cae4f27cc92feb60b104415acb29323ca