Don't mark SAT/DT as failures
Since SAT/DT is not supported in our case, they should be marked as
ignored but in our case they are marked as failures.
Since we don't have SAT/DT support we are returning INVALID_ARGUMENTS
and test is calling the printSkipped function.
Instead of calling printSkipped if we call GTEST_Skip()
unsupported TC results are parsed properly and test cases wil move
to ignored category

Test: VtsHalBroadcastradioV1_0TargetTest
Bug: 191229970

Signed-off-by: Mirche,Dinesh <dinesh.mirche@intel.com>
Change-Id: Iaab0e0ad8c6b22b62b149dc6b24d7d1db89c426e