Wait some time for resetting back to sim power on and then verify it
When sim card power-on request is responsed,
VTS checks the card status immediately.
But card status is not updated at this time.
There is about a 700ms delay here.
There is a 10 seconds timeout to check sim card status.
Adding sleep(1) to check sim card status when sim card is absent
after power-on

BUG: b/204158099

Test: attest VtsHalRadioV1_6TargetTest

Change-Id: I9704f39b7678bdb351c4e700033161e5c2c107a4
Signed-off-by: Caijiao Zhang<zhangcj1@motorola.com>
Signed-off-by: Yinchu Chen<chenyc5@motorola.com>