send <request active source> command to test cec message
VtsHalTvCec uses <give physical address> command with broadcast mode to test
cec sendMessage function, however according to the cec spec about the command's
description which says "A request to a device to return its Physical Address",
so this command is not right for broadcast, some cec drivers deny to broadcast
this command, which lead the VTS case fail.

So I suggestion to use <request active source> command to do the job.

Change-Id: I040c72180cf90cfbceb6363a50a621acf888c654