Add UsageSetting to IRadioNetwork
Add usage setting to IRadioNetwork(Response) and the associated
enum.

Expose the cellular usage setting to Android to be able to query and
set the modem to operate in a data-centric mode or a voice-centricm
mode.
This facilitates "companion" subscriptions that might be data-only and
allows voice-centric devices to operate in data-centric mode on networks
that do not support voice service.

Bug: 210023167
Test: compilation
Change-Id: I61b559589a51d139374dc52599201b6412d93da5