Add Response when set/getUsageSetting Unsupported
Setting and Getting the cellular Usage Setting is only
supportd on AIDL (radio HAL 2.0). While Android should never
actually invoke these methods on older HALs, if one is invoked
there is no response, which means that the RIL will have to wait
for a timeout on these messages. Instead, a response error
can be sent indicating that the capability is unsupported and
unblocking the RIL.

Bug: 212592515
Test: make android.hardware.radio-library.compat
Change-Id: I5b7ecd1acd150fc006430b925ce750156176bf5e